requirements.txt
# Packages requirements for updating the requirements.txt.freeze list via
#  make update-packages
# Deployments will use requirements.txt.freeze which is generated by make update-packages

kafka-python>=2.0
psycopg2-binary>=2.8.0

# commandline interface
click>=7.1

# scraping and pinging
# Usually I would use requests, but this seemed like a nice time to try httpx out, which seems to be the
# successor due to asyncio support: https://github.com/encode/httpx
# -> "considered in beta [...] API stable [...] A 1.0 release is expected to be issued sometime in late 2020."
httpx>=0.16.0<0.17

# testing
pytest>=6.1

# for downloading certs and configuring the "deplyoed" version of kafka/postgresql
aiven-client
